refactor: Separate decision of bar dimension
